separeted the category selector into a diffetent modal
