+++ app/src/pages/HelpPages/CreateHelpOffer/styles.js
@@ -52,22 +52,6 @@ const styles = StyleSheet.create({
     descriptionInput: {
         marginTop: 20,
     },
-    addCategory: {
-        backgroundColor: colors.primary,
-        borderWidth: 1,
-        borderColor: colors.primary,
-        padding: 5,
-        marginTop: 20,
-        width: '40%',
-        borderRadius: 5,
-        justifyContent: 'center',
-        alignItems: 'center',
-    },
-    addCategoryText: {
-        ...fonts.body,
-        color: '#fff',
-        fontSize: 18,
-    },
 });
 
 export default styles;
